+++ src/store.js
@@ -1,19 +1,10 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+import { configureStore } from '@reduxjs/toolkit';
+import randomMessageSlice from './randomMessageSlice';
 
-export const fetchMessage = createAsyncThunk('message/fetchMessage', async () => {
-  const response = await fetch('http://localhost:3000/api/v1/messages/random');
-  const data = await response.json();
-  return data.greeting;
-});
-const messageSlice = createSlice({
-  name: 'message',
-  initialState: { message: '' },
-  reducers: {},
-  extraReducers: (builder) => {
-    builder.addCase(fetchMessage.fulfilled,
-      (state, action) => ({ ...state, message: action.payload }
-      ));
+const store = configureStore({
+  reducer: {
+    randomMessage: randomMessageSlice.reducer,
   },
 });
-export const selectMessage = (state) => state.message.message;
-export default messageSlice.reducer;
+
+export default store;
